fix: give access to the SITE_DOMAIN env var in runtime
